enhancements(rafty): Put vars and types into scoped type files
See CONTRIBUTING.md for more info
