+++ CONTRIBUTING.md
@@ -22,3 +22,8 @@ To delete them:
 for i in $(seq 2 10); do sudo ifconfig lo0 delete 127.0.0.$i;done
 
 ```
+
+## Types
+
+In `golang`, it's pretty common to put vars and types at the beginning of the file. By the time, the file keep growing and at some point, you will see in some projects those definitions can be found in the middle of nowhere. Let's avoid that pattern and put everything in scoped type files. As an example for rafty scope, it will be `rafty_types.go`.\
+This will make this project more consistant by providing a better clarity of what is being done.
