+++ rpcs.go
@@ -9,133 +9,6 @@ import (
 	"google.golang.org/grpc"
 )
 
-// RPCType is used to build rpc requests
-type RPCType uint8
-
-const (
-	// AskNodeID will be used to ask node id
-	AskNodeID RPCType = iota
-
-	// GetLeader will be used to ask who is the leader
-	GetLeader
-
-	// PreVoteRequest is used during pre vote request
-	PreVoteRequest
-
-	// VoteRequest is used during Vote request
-	VoteRequest
-
-	// AppendEntryRequest is used during append entry request
-	AppendEntryRequest
-
-	// ForwardCommandToLeader is used to forward command to leader
-	ForwardCommandToLeader
-
-	// TimeoutNow is used during leadership transfer
-	TimeoutNowRequest
-
-	// MembershipChangeRequest is used during membership changes
-	// by a new node that wants to be part of the cluster
-	MembershipChangeRequest
-
-	// BootstrapClusterRequest is used to bootstrap the cluster
-	BootstrapClusterRequest
-
-	// InstallSnapshotRequest is used to install snapshot in the cluster
-	InstallSnapshotRequest
-)
-
-// RPCRequest is used by chans in order to manage rpc requests
-type RPCRequest struct {
-	RPCType      RPCType
-	Request      any
-	Timeout      time.Duration
-	ResponseChan chan<- RPCResponse
-}
-
-// RPCResponse  is used by RPCRequest in order to reply to rpc requests
-type RPCResponse struct {
-	TargetPeer Peer
-	Response   any
-	Error      error
-}
-
-// RPCAskNodeIDRequest hold the requirements to ask node id
-type RPCAskNodeIDRequest struct {
-	Id, Address string
-}
-
-// RPCAskNodeIDResponse hold the response from RPCAskNodeIDRequest
-type RPCAskNodeIDResponse struct {
-	LeaderID, LeaderAddress, PeerID string
-	ReadReplica, AskForMembership   bool
-}
-
-// RPCGetLeaderRequest hold the requirements to get the leader
-type RPCGetLeaderRequest struct {
-	PeerID, PeerAddress string
-	TotalPeers          int
-}
-
-// RPCGetLeaderResponse hold the response from RPCGetLeaderRequest
-type RPCGetLeaderResponse struct {
-	LeaderID, LeaderAddress, PeerID string
-	TotalPeers                      int
-	AskForMembership                bool
-}
-
-// RPCPreVoteRequest hold the requirements to send pre vote requests
-type RPCPreVoteRequest struct {
-	Id          string
-	CurrentTerm uint64
-}
-
-// RPCPreVoteResponse hold the response from RPCPreVoteRequest
-type RPCPreVoteResponse struct {
-	PeerID                     string
-	RequesterTerm, CurrentTerm uint64
-	Granted                    bool
-}
-
-// RPCVoteRequest hold the requirements to send vote requests
-type RPCVoteRequest struct {
-	CandidateId, CandidateAddress          string
-	CurrentTerm, LastLogIndex, LastLogTerm uint64
-	CandidateForLeadershipTransfer         bool
-}
-
-// RPCVoteResponse hold the response from RPCVoteRequest
-type RPCVoteResponse struct {
-	PeerID                     string
-	RequesterTerm, CurrentTerm uint64
-	Granted                    bool
-}
-
-// RPCTimeoutNowRequest hold the requirements to send timeout now requests
-// for leadership transfer
-type RPCTimeoutNowRequest struct{}
-
-// RPCTimeoutNowResponse hold the response from RPCTimeoutNowRequest
-type RPCTimeoutNowResponse struct {
-	Success bool
-}
-
-// RPCMembershipChangeRequest hold the requirements to send membership requests
-type RPCMembershipChangeRequest struct {
-	Address, Id               string
-	ReadReplica               bool
-	Action                    uint32
-	LastLogIndex, LastLogTerm uint64
-}
-
-// RPCMembershipChangeResponse hold the response from RPCMembershipChangeRequest
-type RPCMembershipChangeResponse struct {
-	ActionPerformed, Response uint32
-	LeaderID, LeaderAddress   string
-	Peers                     []Peer
-	Success                   bool
-}
-
 // sendRPC is use to send rpc request
 func (r *Rafty) sendRPC(request RPCRequest, client raftypb.RaftyClient, peer Peer) {
 	options := []grpc.CallOption{}
