+++ snapshot_interval_types.go
@@ -1,0 +1,4 @@
+package rafty
+
+// snapshotTestHook is used for injecting errors during unit testing
+var snapshotTestHook func() error
