+++ state_follower.go
@@ -5,12 +5,6 @@ import (
 	"time"
 )
 
-// follower hold all requirements by a node in follower state
-type follower struct {
-	// rafty holds rafty config
-	rafty *Rafty
-}
-
 // init initialize all requirements needed by
 // the current node type
 func (r *follower) init() {
