+++ state_follower_types.go
@@ -1,0 +1,7 @@
+package rafty
+
+// follower hold all requirements by a node in follower state
+type follower struct {
+	// rafty holds rafty config
+	rafty *Rafty
+}
