+++ state_machine_types.go
