+++ state_read_replica_types.go
@@ -1,0 +1,7 @@
+package rafty
+
+// readReplica hold all requirements by a node in read replica state
+type readReplica struct {
+	// rafty holds rafty config
+	rafty *Rafty
+}
